refactor: move FaceGeometry implementations
