--- Packages/com.github.homuler.mediapipe/Runtime/Scripts/Protobuf/Modules/FaceGeometry/EffectRendererCalculator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 64856f22fa5ced2cdabfde7cf22602e4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
