--- Packages/com.github.homuler.mediapipe/Runtime/Scripts/Protobuf/Modules/FaceGeometry/GeometryPipelineCalculator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 179d060f02ec2c10fb0b66d898be90da
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
