--- Packages/com.github.homuler.mediapipe/Runtime/Scripts/Protobuf/Modules/FaceGeometry/Protos/Environment.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 27261b8f2120b83ca871805186fc2688
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
