+++ Packages/com.github.homuler.mediapipe/Runtime/Scripts/Protobuf/Tasks/Vision/FaceGeometry.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b3540aec19d459a958641688263d096b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
