+++ Packages/com.github.homuler.mediapipe/Runtime/Scripts/Protobuf/Tasks/Vision/FaceGeometry/GeometryPipelineCalculator.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 26bc289e136a1adfcb6608d60a732931
+guid: 4c928d81ce84fc883baf92ad86583f94
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
