+++ Packages/com.github.homuler.mediapipe/Runtime/Scripts/Protobuf/Tasks/Vision/FaceGeometry/Proto.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 611d5da7812ec1a40b0dbc8c80100f8f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
