+++ Packages/com.github.homuler.mediapipe/Runtime/Scripts/Protobuf/Tasks/Vision/FaceGeometry/Proto/FaceGeometry.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4ee83bbf21893b0698762553487f5dba
+guid: fbc44c4232705405f9e85512b4d4dd62
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
