+++ Packages/com.github.homuler.mediapipe/Runtime/Scripts/Protobuf/Tasks/Vision/FaceGeometry/Proto/FaceGeometryGraphOptions.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5c7348ed702634ae1bed9d24f5df5af0
+guid: 162297c24d908d18f87000716acb08f6
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
