+++ Packages/com.github.homuler.mediapipe/Runtime/Scripts/Protobuf/Tasks/Vision/FaceGeometry/Proto/Mesh3D.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fe56866aa90e25df2b520b254aa7d971
+guid: 4e810e14cc52e74d4b388fe270f0f32b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
