+++ Packages/com.github.homuler.mediapipe/Runtime/Scripts/Tasks/Vision/FaceGeometry.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0732d100f034703e88c3cf1d26e973f3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
