+++ Packages/com.github.homuler.mediapipe/Runtime/Scripts/Tasks/Vision/FaceGeometry/FaceGeometryPacket.cs
@@ -1,4 +1,4 @@
-// Copyright (c) 2021 homuler
+// Copyright (c) 2023 homuler
 //
 // Use of this source code is governed by an MIT-style
 // license that can be found in the LICENSE file or at
@@ -6,9 +6,9 @@
 
 using System;
 
-namespace Mediapipe
+namespace Mediapipe.Tasks.Vision.FaceGeometry
 {
-  public class FaceGeometryPacket : Packet<FaceGeometry.FaceGeometry>
+  public class FaceGeometryPacket : Packet<Proto.FaceGeometry>
   {
     /// <summary>
     ///   Creates an empty <see cref="FaceGeometryPacket" /> instance.
@@ -18,25 +18,19 @@ namespace Mediapipe
     [UnityEngine.Scripting.Preserve]
     public FaceGeometryPacket(IntPtr ptr, bool isOwner = true) : base(ptr, isOwner) { }
 
-    public FaceGeometryPacket At(Timestamp timestamp)
-    {
-      return At<FaceGeometryPacket>(timestamp);
-    }
+    public FaceGeometryPacket At(Timestamp timestamp) => At<FaceGeometryPacket>(timestamp);
 
-    public override FaceGeometry.FaceGeometry Get()
+    public override Proto.FaceGeometry Get()
     {
       UnsafeNativeMethods.mp_Packet__GetFaceGeometry(mpPtr, out var serializedProto).Assert();
       GC.KeepAlive(this);
 
-      var geometry = serializedProto.Deserialize(FaceGeometry.FaceGeometry.Parser);
+      var geometry = serializedProto.Deserialize(Proto.FaceGeometry.Parser);
       serializedProto.Dispose();
 
       return geometry;
     }
 
-    public override FaceGeometry.FaceGeometry Consume()
-    {
-      throw new NotSupportedException();
-    }
+    public override Proto.FaceGeometry Consume() => throw new NotSupportedException();
   }
 }
