+++ Packages/com.github.homuler.mediapipe/Runtime/Scripts/Tasks/Vision/FaceGeometry/FaceGeometryPacket.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f95c0ec47a36687ba8e49a70c3261e78
+guid: 75c38d6d8b20e1f0fa3623f352e257cb
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
