+++ Packages/com.github.homuler.mediapipe/Runtime/Scripts/Tasks/Vision/FaceGeometry/FaceGeometryVectorPacket.cs
@@ -1,4 +1,4 @@
-// Copyright (c) 2021 homuler
+// Copyright (c) 2023 homuler
 //
 // Use of this source code is governed by an MIT-style
 // license that can be found in the LICENSE file or at
@@ -7,9 +7,9 @@
 using System;
 using System.Collections.Generic;
 
-namespace Mediapipe
+namespace Mediapipe.Tasks.Vision.FaceGeometry
 {
-  public class FaceGeometryVectorPacket : Packet<List<FaceGeometry.FaceGeometry>>
+  public class FaceGeometryVectorPacket : Packet<List<Proto.FaceGeometry>>
   {
     /// <summary>
     ///   Creates an empty <see cref="FaceGeometryVectorPacket" /> instance.
@@ -19,25 +19,19 @@ namespace Mediapipe
     [UnityEngine.Scripting.Preserve]
     public FaceGeometryVectorPacket(IntPtr ptr, bool isOwner = true) : base(ptr, isOwner) { }
 
-    public FaceGeometryVectorPacket At(Timestamp timestamp)
-    {
-      return At<FaceGeometryVectorPacket>(timestamp);
-    }
+    public FaceGeometryVectorPacket At(Timestamp timestamp) => At<FaceGeometryVectorPacket>(timestamp);
 
-    public override List<FaceGeometry.FaceGeometry> Get()
+    public override List<Proto.FaceGeometry> Get()
     {
       UnsafeNativeMethods.mp_Packet__GetFaceGeometryVector(mpPtr, out var serializedProtoVector).Assert();
       GC.KeepAlive(this);
 
-      var geometries = serializedProtoVector.Deserialize(FaceGeometry.FaceGeometry.Parser);
+      var geometries = serializedProtoVector.Deserialize(Proto.FaceGeometry.Parser);
       serializedProtoVector.Dispose();
 
       return geometries;
     }
 
-    public override List<FaceGeometry.FaceGeometry> Consume()
-    {
-      throw new NotSupportedException();
-    }
+    public override List<Proto.FaceGeometry> Consume() => throw new NotSupportedException();
   }
 }
