+++ Packages/com.github.homuler.mediapipe/Runtime/Scripts/Tasks/Vision/FaceGeometry/FaceGeometryVectorPacket.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 263876dc23a587949869ee40a64e2988
+guid: 877885ba96b5ca458816b28feec7a3de
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
