+++ mediapipe_api/BUILD
@@ -164,7 +164,7 @@ cc_library(
         "//mediapipe_api/framework/formats:matrix_data",
         "//mediapipe_api/framework/formats:rect",
         "//mediapipe_api/framework/port:logging",
-        "//mediapipe_api/modules/face_geometry/protos:face_geometry",
+        "//mediapipe_api/tasks/cc/vision/face_geometry/proto:face_geometry",
         "//mediapipe_api/tasks/cc/core:task_runner",
         "//mediapipe_api/util:resource_util",
     ] + select({
@@ -560,11 +560,11 @@ pkg_zip(
         "//mediapipe_api/framework/formats/object_detection:proto_srcs",
         "//mediapipe_api/gpu:proto_srcs",
         "//mediapipe_api/modules/face_detection:proto_srcs",
-        "//mediapipe_api/modules/face_geometry:proto_srcs",
-        "//mediapipe_api/modules/face_geometry/protos:proto_srcs",
         "//mediapipe_api/modules/holistic_landmark/calculators:proto_srcs",
         "//mediapipe_api/tasks/cc/core/proto:proto_srcs",
         "//mediapipe_api/tasks/cc/vision/face_detector/proto:proto_srcs",
+        "//mediapipe_api/tasks/cc/vision/face_geometry/proto:proto_srcs",
+        "//mediapipe_api/tasks/cc/vision/face_geometry/calculators:proto_srcs",
         "//mediapipe_api/util:proto_srcs",
         "//mediapipe_api/util/tracking:proto_srcs",
     ],
